moved pages into pages folder, standardized capitalization
