+++ packages/express_backend/routes/grocery-routes.ts
@@ -65,23 +65,28 @@ groceryRouter.delete(
 	}
 );
 
-groceryRouter.put("/:home/grocery/:id", async (req: Request, res: Response) => {
-	try {
-		const updated = await updateGroceryItem(req.params.id, req.body);
+groceryRouter.patch(
+	"/:home/grocery/:id",
+	async (req: Request, res: Response) => {
+		try {
+			const updated = await updateGroceryItem(req.params.id, req.body);
 
-		if (!updated) {
-			return res.status(404).json({ error: "Grocery item not found" });
+			if (!updated) {
+				return res
+					.status(404)
+					.json({ error: "Grocery item not found" });
+			}
+
+			res.status(200).json(updated);
+		} catch (error) {
+			console.error(error);
+
+			res.status(400).json({ error: "Invalid ID" });
 		}
-
-		res.status(200).json(updated);
-	} catch (error) {
-		console.error(error);
-
-		res.status(400).json({ error: "Invalid ID" });
 	}
-});
+);
 
-groceryRouter.put(
+groceryRouter.patch(
 	"/:home/grocery/:id/:newQuantity",
 	async (req: Request, res: Response) => {
 		try {
