+++ packages/react-frontend/src/pages/rulesPage.tsx
